duel bug fixes and implemented removed from prototype
