+++ src/main/java/lollipop/commands/duel/models/DCPUAI.java
@@ -1,0 +1,7 @@
+package lollipop.commands.duel.models;
+
+public class DCPUAI {
+
+
+
+}
